cookieに保存できるようにした。　axiosでフロントエンドでAPIを叩けるようした。　次はフロントエンドの仕組みを書く

+++ back_end/mysite/goods/views.py
@@ -2,7 +2,8 @@ from django.shortcuts import render
 from rest_framework import viewsets
 from .models import Goods
 from .serializers import GoodsSerializer
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework import status
 from django.shortcuts import get_object_or_404
@@ -21,27 +22,39 @@ def getAllGoods(request):
     return Response(serializer.data)
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def createGoods(request):
-    data = request.data
-    
-    good = Goods.objects.create(
-        title = data['title'],
-        content = data['content'],
-        picture = data['picture'],
-    )
+    user = request.user
+    print(user)
+    if user.is_superuser:
+        data = request.data
+        
+        good = Goods.objects.create(
+            title = data['title'],
+            content = data['content'],
+            picture = data['picture'],
+        )
+    else:
+        Response({"messege": "admin only access"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)    
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def updateGood(request, pk):
-    data = request.data
-    good = Goods.objects.get(id=pk)
-    
-    good.title = data['title']
-    good.content = data['content']
-    good.picture = data['picture']
-    good.public = data['public']
-    
-    good.save()
-    serializer = GoodsSerializer(good, many=False)
-    return Response(serializer.data)
+    user = request.user
+    if user.is_superuser:
+        data = request.data
+        good = get_object_or_404(Goods, id=pk)
+        
+        good.title = data['title']
+        good.content = data['content']
+        good.picture = data['picture']
+        good.public = data['public']
+        
+        good.save()
+        serializer = GoodsSerializer(good, many=False)
+        return Response(serializer.data)
+    else:
+        Response({"messege": "admin only access"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+        
 
 @api_view(['GET'])
 def getGoodsById(request, pk):
@@ -52,9 +65,19 @@ def getGoodsById(request, pk):
         return Response(serializer.data)
     except:
         return Response({"error": "error..."}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+
+
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def deleteGood(request, pk):
-    good = Goods.objects.get(id=pk)
-    good.delete()
+    user = request.user
+    if user.is_superuser:
     
-    return Response({'messege':'delete success'})
+        good = get_object_or_404(Goods, id=pk)
+        
+        good.delete()
+        
+        return Response({'messege':'delete success'})
+
+    else:
+        Response({"messege": "admin only access"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
